+++ README.md
@@ -18,3 +18,6 @@
 - Форма добавления видов продуктов (новая таблица, в ней названия типа "хлеб")
 - Форма добавления комментария (выбираешь продукт из тех, что в холодильнике и пишешь коммент)
 - На форме "я купил это" добавить выбор категории продукта ("Маасдам ворованный" - это "сыр")
+
+## А вот на этой форме чеклист с проблемами, которые надо проверять
+...
