+++ Холодильник/Admin.cs
@@ -70,5 +70,15 @@ namespace Fridgerator
         {
 
         }
+
+        private void Admin_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void dateTimePicker1_ValueChanged(object sender, EventArgs e)
+        {
+
+        }
     }
 }
