+++ Холодильник/Contents.cs
@@ -93,5 +93,10 @@ namespace Fridgerator
 
             UpdateTable(products);
         }
+
+        private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        {
+
+        }
     }
 }
